# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
      "AIModel": {
        "description": "Schema for AI model information",
        "properties": {
          "display_name": {
            "description": "Human-readable model name",
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "is_active": {
            "default": true,
            "description": "Whether the model is active",
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "description": "Model identifier (e.g., 'gemini-2.0-flash')",
            "title": "Name",
            "type": "string"
          },
          "provider_id": {
            "description": "Reference to AI provider",
            "title": "Provider Id",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Model version (e.g., '001', 'latest')",
            "title": "Version"
          }
        },
        "required": [
          "provider_id",
          "name",
          "display_name"
        ],
        "title": "AIModel",
        "type": "object"
      },
...
      "GeminiConfigRequest": {
        "description": "Schema for creating/updating Gemini configuration",
        "properties": {
          "api_key": {
            "description": "Gemini API key",
            "title": "Api Key",
            "type": "string"
          },
          "default_model_name": {
            "description": "Default model name (e.g., 'gemini-2.0-flash')",
            "title": "Default Model Name",
            "type": "string"
          }
        },
        "required": [
          "api_key",
          "default_model_name"
        ],
        "title": "GeminiConfigRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UserAIConfigDisplay": {
        "description": "Schema for displaying user AI config (with masked API key)",
        "properties": {
          "api_key_masked": {
            "description": "Masked API key (e.g., 'AIza****')",
            "title": "Api Key Masked",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "default_model_display_name": {
            "description": "Human-readable model name",
            "title": "Default Model Display Name",
            "type": "string"
          },
          "default_model_name": {
            "description": "Default model name",
            "title": "Default Model Name",
            "type": "string"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "provider_display_name": {
            "description": "Human-readable provider name",
            "title": "Provider Display Name",
            "type": "string"
          },
          "provider_name": {
            "description": "Provider name (e.g., 'gemini')",
            "title": "Provider Name",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "provider_name",
          "provider_display_name",
          "api_key_masked",
          "default_model_name",
          "default_model_display_name"
        ],
        "title": "UserAIConfigDisplay",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Graphora API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/ai-config": {
      "get": {
        "description": "Get user's AI configuration\n\nArgs:\n    user_id: User's ID (from header)\n\nReturns:\n    UserAIConfigDisplay: User's AI configuration with masked API key\n\nRaises:\n    HTTPException: 404 if configuration not found",
        "operationId": "get_user_ai_config_api_v1_ai_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAIConfigDisplay"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get User Ai Config",
        "tags": [
          "AI Configuration"
        ]
      }
    },
    "/api/v1/ai-config/gemini": {
      "post": {
        "description": "Create a new Gemini configuration for a user\n\nArgs:\n    config_request: Gemini configuration data\n    user_id: User's ID (from header)\n\nReturns:\n    UserAIConfigDisplay: Created configuration with masked API key\n\nRaises:\n    HTTPException: 400 if configuration already exists or validation fails",
        "operationId": "create_gemini_config_api_v1_ai_config_gemini_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeminiConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAIConfigDisplay"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Gemini Config",
        "tags": [
          "AI Configuration"
        ]
      },
      "put": {
        "description": "Update an existing Gemini configuration for a user\n\nArgs:\n    config_request: Updated Gemini configuration data\n    user_id: User's ID (from header)\n\nReturns:\n    UserAIConfigDisplay: Updated configuration with masked API key\n\nRaises:\n    HTTPException: 404 if configuration not found, 400 if validation fails",
        "operationId": "update_gemini_config_api_v1_ai_config_gemini_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeminiConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAIConfigDisplay"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Gemini Config",
        "tags": [
          "AI Configuration"
        ]
      }
    },
    "/api/v1/ai-models/{provider_name}": {
      "get": {
        "description": "Get all available models for a specific AI provider\n\nArgs:\n    provider_name: Name of the AI provider (e.g., 'gemini')\n\nReturns:\n    List[AIModel]: List of available models for the provider",
        "operationId": "get_ai_models_api_v1_ai_models__provider_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "provider_name",
            "required": true,
            "schema": {
              "title": "Provider Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AIModel"
                  },
                  "title": "Response Get Ai Models Api V1 Ai Models  Provider Name  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ai Models",
        "tags": [
          "AI Configuration"
        ]
      }
    },
...
  }
}